+++ database/migrations/2019_01_24_154848_create_fan_tickets_table.php
@@ -15,11 +15,11 @@ class CreateFanTicketsTable extends Migration
     {
         Schema::create('fan_tickets', function (Blueprint $table) {
             $table->increments('id');
-            $table->tinyInteger('fan_id')->commend('粉丝id');
-            $table->tinyInteger('ticket_id')->commend('门票id');
+            $table->integer('fan_id')->commend('粉丝id');
+            $table->integer('ticket_id')->commend('门票id');
             $table->string('name', 20)->default('')->comment('真实姓名');
             $table->string('mobile')->default('')->comment('手机号码');
-            $table->tinyInteger('purchase_quantity')->comment('购买数量');
+            $table->integer('purchase_quantity')->comment('购买数量');
             $table->dateTime('booking_date')->comment('预约日期');
             $table->timestamps();
         });
